requirements.txt
streamlit==1.32.0
plotly==5.19.0
pandas==2.2.0
numpy==1.26.0
scipy==1.12.0
pymc==5.10.0
statsmodels==0.14.1
lifelines==0.27.8
dbt-core==1.5.11
dbt-sqlite==1.5.1
google-generativeai==0.4.1
python-dotenv==1.0.0
pytest==8.0.0
openpyxl==3.1.2
pytensor==2.18.6
arviz==0.17.0
xarray==2024.1.0
cloudpickle==3.0.0
